requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
daff==1.4.2
db-dtypes==1.4.3
dbt-adapters==1.17.2
dbt-bigquery==1.10.2
dbt-common==1.32.0
dbt-core==1.11.0b3
dbt-extractor==0.6.0
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.1
docstring_parser==0.17.0
duckdb==1.4.1
exceptiongroup==1.3.0
fastjsonschema==2.21.2
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.2
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.119.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-dataproc==5.22.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.41.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
leather==0.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.7.0
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
orderly-set==5.5.0
packaging==24.2
pandas==2.3.3
pandas-gbq==0.29.2
parsedatetime==2.6
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
sqlparse==0.5.3
streamlit==1.50.0
tenacity==8.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
zipp==3.23.0
matplotlib>=3.8,<3.11
scipy>=1.10
fastapi==0.115.2
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
xgboost==2.1.2
shap==0.46.0
